ui/progress: extract signal handling into own package
